requirements.txt
# Dependencies for federal ingestion
requests==2.31.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
tenacity==8.2.3
structlog==23.2.0
python-dotenv==1.0.0
pydantic[email]==1.10.15
redis==5.0.1
